open file channel only once; make LockedFileChannel internal
